# Multithread chromosome