--- a/Haplotag.cpp
+++ b/Haplotag.cpp
@@ -22,12 +22,13 @@ static const char *CORRECT_USAGE_MESSAGE =
 "                                      if the alignment has no obvious corresponding haplotype, it will not be tagged. default:0.6\n"
 "      -t, --threads=Num               number of thread. default:1\n"
 "      -o, --out-prefix=NAME           prefix of phasing result. default:result\n"
+"      --region=REGION                 tagging include only reads/variants overlapping those regions. default:""(all regions)"
 "      --log                           an additional log file records the result of each read. default:false\n";
 
 
 static const char* shortopts = "s:b:o:t:q:p:r:";
 
-enum { OPT_HELP = 1, TAG_SUP, SV_FILE, LOG, MOD_FILE};
+enum { OPT_HELP = 1, TAG_SUP, SV_FILE, REGION, LOG, MOD_FILE};
 
 static const struct option longopts[] = { 
     { "help",                 no_argument,        NULL, OPT_HELP },
@@ -41,6 +42,7 @@ static const struct option longopts[] = {
     { "threads",              required_argument,  NULL, 't' },
     { "qualityThreshold",     required_argument,  NULL, 'q' },
     { "percentageThreshold",  required_argument,  NULL, 'p' },
+    { "region",               required_argument,  NULL, REGION },
     { "log",                  no_argument,        NULL, LOG },
     { NULL, 0, NULL, 0 }
 };
@@ -56,6 +58,7 @@ namespace opt
     static std::string bamFile="";
     static std::string fastaFile="";
     static std::string resultPrefix="result";
+    static std::string region="";
     static bool tagSupplementary = false;
     static bool writeReadLog = false;
 }
@@ -80,6 +83,7 @@ void HaplotagOptions(int argc, char** argv)
         case SV_FILE: arg >> opt::svFile; break;
         case MOD_FILE: arg >> opt::modFile; break;
         case TAG_SUP: opt::tagSupplementary = true; break;
+        case REGION: arg >> opt::region; break;
         case LOG: opt::writeReadLog = true; break;
         case OPT_HELP:
             std::cout << CORRECT_USAGE_MESSAGE;
@@ -193,6 +197,7 @@ int HaplotagMain(int argc, char** argv)
     ecParams.resultPrefix=opt::resultPrefix;
     ecParams.tagSupplementary=opt::tagSupplementary;
     ecParams.percentageThreshold=opt::percentageThreshold;
+    ecParams.region=opt::region;
     ecParams.writeReadLog=opt::writeReadLog;
 
     HaplotagProcess processor(ecParams);