--- a/ParsingBam.cpp
+++ b/ParsingBam.cpp
@@ -11,21 +11,21 @@
 
 
 // FASTA
-FastaParser::FastaParser(std::string fastaFile ,  std::vector<std::string> chrName , std::vector<int> last_pos):
+FastaParser::FastaParser(std::string fastaFile,  std::vector<std::string> chrName, std::vector<int> last_pos, int numThreads):
     fastaFile(fastaFile),
     chrName(chrName),
     last_pos(last_pos)
 {
-    if(fastaFile==""){
-        for(std::vector<std::string>::iterator iter = chrName.begin() ; iter != chrName.end() ; iter++)
-            chrString.insert(std::make_pair( (*iter) , ""));
-        return;
-    }
-    
-    faidx_t *fai = NULL;
-    fai = fai_load(fastaFile.c_str());
+    // init map
+    for(std::vector<std::string>::iterator iter = chrName.begin() ; iter != chrName.end() ; iter++)
+        chrString.insert(std::make_pair( (*iter) , ""));
+
     // iterating all chr
+    #pragma omp parallel for schedule(dynamic) num_threads(numThreads)
     for(std::vector<std::string>::iterator iter = chrName.begin() ; iter != chrName.end() ; iter++){
+        
+        faidx_t *fai = NULL;
+        fai = fai_load(fastaFile.c_str());
         int index = iter - chrName.begin();
 
         // ref_len is a return value that is length of retrun string
@@ -35,8 +35,8 @@ FastaParser::FastaParser(std::string fastaFile ,  std::vector<std::string> chrNa
         if(ref_len == 0){
             std::cout<<"nothing in reference file \n";
         }
-        // insert to map
-        chrString.insert(std::make_pair( (*iter) , chr_info));
+        // update map
+        chrString[(*iter)] = chr_info;
 
     }
 }
@@ -923,8 +923,8 @@ void BamParser::direct_detect_alleles(int lastSNPPos, int &numThreads, PhasingPa
         firstVariantIter = tmpFirstVariantIter;
         firstSVIter = tmpFirstSVIter;
         firstModIter = tmpFirstModIter;
-              
-        // init data structure and get core n
+
+// init data structure and get core n
         htsThreadPool threadPool = {NULL, 0};
         // open bam file
         samFile *fp_in = hts_open(bamFile.c_str(),"r"); 
@@ -945,8 +945,8 @@ void BamParser::direct_detect_alleles(int lastSNPPos, int &numThreads, PhasingPa
         hts_itr_t* iter = sam_itr_querys(idx, bamHdr, range.c_str());
 
         
-        int result;
-        
+                int result;
+
         // creat thread pool
         if (!(threadPool.pool = hts_tpool_init(numThreads))) {
             fprintf(stderr, "Error creating thread pool\n");
@@ -973,7 +973,7 @@ void BamParser::direct_detect_alleles(int lastSNPPos, int &numThreads, PhasingPa
         bam_hdr_destroy(bamHdr);
         bam_destroy1(aln);
         sam_close(fp_in);
-        hts_tpool_destroy(threadPool.pool);
+hts_tpool_destroy(threadPool.pool);
     }
     
 }