--- a/Phasing.cpp
+++ b/Phasing.cpp
@@ -18,8 +18,7 @@ static const char *CORRECT_USAGE_MESSAGE =
 "optional arguments:\n"
 "   --sv-file=NAME                         input SV vcf file.\n"
 "   --mod-file=NAME                        input modified vcf file.(produce by longphase modcall)\n"
-"   -t, --threads=Num                      number of thread. default:0\n"
-"                                          If set to 0, use all available threads.\n"
+"   -t, --threads=Num                      number of thread. default:1\n"
 "   -o, --out-prefix=NAME                  prefix of phasing result. default: result\n"
 "   --indels                               phase small indel. default: False\n"
 "   --dot                                  each contig/chromosome will generate dot file. \n\n"
@@ -68,7 +67,7 @@ static const struct option longopts[] = {
 
 namespace opt
 {
-    static int numThreads = 0;
+    static int numThreads = 1;
     static int distance = 300000;
     static std::string snpFile="";
     static std::string svFile="";
@@ -179,7 +178,7 @@ void PhasingOptions(int argc, char** argv)
         die = true;
     }  
     
-    if ( opt::numThreads < 0 ){
+    if ( opt::numThreads < 1 ){
         std::cerr << SUBPROGRAM " invalid threads. value: " 
                   << opt::numThreads 
                   << "\n please check -t, --threads=Num\n";