- commit : [multithread fastaParser](https://github.com/twolinin/longphase/commit/2844e8a74dba6745b55f81d910f4039d5a64683b)
  data : GCA_000001405.15_GRCh38_no_alt_analysis_set.fa
  threads : 24
  before : 21s
  after : 2s
  use the feature : phase, haplotag
- commit : [thread pool to manage thread allocation](https://github.com/twolinin/longphase/commit/85c15d189d2a377cffc8460e914b268096dbeb4a)
  use the feature : phase, modcall

  This test compares the run times with the develop branch (commit https://github.com/twolinin/longphase/commit/d978d221f1cc1d1628cb214d47af35293133b326), performing phasing at 10x to 60x scale. The time format is `mm:ss`.
  Test Condition threads =24 | Run Time Before Optimization | Run Time After Optimization
  -- | -- | --
  HG002 ONT 10x | 01:22.77 | 00:53.60
  HG002 ONT 20x | 02:06.08 | 01:36.08
  HG002 ONT 30x | 02:58.30 | 01:26.28
  HG002 ONT 40x | 03:26.56 | 02:23.78
  HG002 ONT 50x | 04:05.03 | 03:15.71
  HG002 ONT 60x | 04:32.25 | 03:40.02
  
  Test Condition threads =100 | Run Time Before Optimization | Run Time After Optimization
  -- | -- | --
  HG002 ONT 10x | 01:26.66 | 00:53.60
  HG002 ONT 20x | 01:59.86 | 01:06.69
  HG002 ONT 30x | 02:42.54 | 01:48.90
  HG002 ONT 40x | 03:42.33 | 02:17.65
  HG002 ONT 50x | 04:01.23 | 02:45.05
  HG002 ONT 60x | 04:20.64 | 02:28.73

- commit : [Correct the conditions for judging the VCF header](https://github.com/twolinin/longphase/commit/2c505d391daab53d102ca5414ebbfd4233bffce3)
  use the feature : haplotag
- commit : [add region functionality](https://github.com/twolinin/longphase/commit/4446ecdb5170f848373d7c32256497bd73a9fb10)
  use the feature : haplotag